requirements.txt
flask
nltk
numpy
scikit-learn
openai
python-dotenv
flask-cors
gunicorn